grpc-server/requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
pytest==7.1.2
pytest-grpc==0.8.0